requirement.txt
pandas 
scikit-learn
numpy
matplotlib
nltk
pickledb
Django
djangorestframework
drf_yasg
django-cors-headers
tensorflow
psycopg2
djangorestframework-simplejwt
corsheaders
rolepermissions
django_filters
django_rest_passwordreset
requests